% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecasts.R
\name{get_forecasts}
\alias{get_forecasts}
\title{Generate forecasts of attendant care hours}
\usage{
get_forecasts(data, h, nsim)
}
\arguments{
\item{data}{Data set computed from \code{\link{read_tac_data}}}

\item{h}{Forecast horizon.}

\item{nsim}{Number of simulated future sample paths per model.}
}
\description{
Generate forecasts from ETS and ARIMA models, reconcile them, and combine them.
Return a fable object containing the forecasts.
}
