% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_costs.R
\name{group_costs}
\alias{group_costs}
\title{Synthetic data for attendant hours by age group and injury group}
\format{
Time series of class `tsibble`
}
\source{
Synthetic data
}
\description{
Artificial Transport Accident Commission attendant care data
}
\details{
\code{group_costs} is a daily `tsibble` with index `billing_period` and 
two values:
\tabular{ll}{
    adjusted_hours:      \tab Total attendant care hours\cr
    nclaims:     \tab Number of active claims \cr
}

The data is disaggregated using two keys:
\tabular{ll}{
    age_group:      \tab Age group of client at the time of accident \cr
    injury_group:      \tab Injury sustained by client due to accident \cr
}
}
\examples{
group_costs

}
\keyword{datasets}
