% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_forecasts}
\alias{plot_forecasts}
\title{Plot forecasts of attendant care hours disaggregated by age or injury group.}
\usage{
plot_forecasts(
  forecasts,
  data,
  show_age_group = "<aggregated>",
  show_injury_group = "<aggregated>"
)
}
\arguments{
\item{forecasts}{A fable object created by \code{\link{get_forecasts}}}

\item{data}{The data used to construct the forecasts. This should be a tsibble object of the same form as \code{\link{group_costs}}.}

\item{show_age_group}{A character string specifying either a specific age group or "<aggregated>" meaning the total across all age groups.}

\item{show_injury_group}{A character string specifying either a specific injury group or "<aggregated>" meaning the total across all injury groups.}
}
\description{
Produce a time plot of attendant care hours per billing period for specific disaggregations.
}
\examples{
\dontrun{
  group_costs |> 
    get_forecasts(h=13, nsim=100) |> 
    plot_forecasts(group_costs)
}
}
\author{
Rob J Hyndman
}
