% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_total_hours}
\alias{plot_total_hours}
\title{Plot attendant care hours disaggregated by age or injury group.}
\usage{
plot_total_hours(group_costs, variable = NULL, include_average = TRUE)
}
\arguments{
\item{group_costs}{A tsibble containing costs optionally split by a variable}

\item{variable}{Name of disaggregation variable. If NULL, aggregated costs are shown}

\item{include_average}{Should the average cost per billing period be shown?}
}
\description{
Produce a time plot of attendant care hours per billing period disaggregated by the given `variable`
}
\examples{
  group_costs |> 
    plot_total_hours(age_group)
}
\author{
Rob J Hyndman
}
