% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tac_data}
\alias{read_tac_data}
\title{Read in TAC data}
\usage{
read_tac_data(claims_file, costs_file)
}
\arguments{
\item{claims_file}{CSV file containing claims header}

\item{costs_file}{CSV file containing attendant hours}
}
\value{
A tsibble object containing total attendant care adjusted hours
for each billing period, disaggregated by age group and injury group.
The column `nclaims` shows the number of "active" claims in each
billing period.
}
\description{
This function takes two csv files as inputs: one containing the claims header
and the other containing the attendant hours. It returns total hours per
age group and injury group by billing period.
}
\examples{
\dontrun{
group_costs <- read_tac_data(
  claims_file = "T086_claim_header.csv",
  costs_file = "T086_attendant_care_hours.csv"
)
}
}
\author{
Rob J Hyndman
}
