% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{tac_accuracy}
\alias{tac_accuracy}
\title{Compute accuracy statistics}
\usage{
tac_accuracy(forecasts, actuals)
}
\arguments{
\item{forecasts}{A fable object with forecasts, usually the output from \code{\link{get_forecasts}}}

\item{actuals}{A tsibble with actual values.  For example, the output from \code{\link{read_tac_data}}}
}
\value{
A tibble with accuracy statistics
}
\description{
Compute accuracy statistics
}
