% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TACforecasting-package}
\alias{TACforecasting}
\alias{TACforecasting-package}
\title{TACforecasting: Forecasting Functions for the Transport Accident Commission}
\description{
Functions to make hierarchical time series forecasts of attendant care hours easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/TAC_forecasting/}
  \item \url{https://github.com/robjhyndman/TAC_forecasting}
  \item Report bugs at \url{https://github.com/robjhyndman/TAC_forecasting/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

}
\keyword{internal}
