% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{tscv_accuracy}
\alias{tscv_accuracy}
\title{Compute forecasts with a rolling origin and return accuracy statistics}
\usage{
tscv_accuracy(group_costs, h, nsim, init, step)
}
\arguments{
\item{group_costs}{A tsibble with actual values. For example, the output from \code{\link{read_tac_data}}}

\item{h}{The forecast horizon}

\item{nsim}{The number of simulations used in each forecast for each model.}

\item{init}{The number of initial observations to use for the first fold.}

\item{step}{The number of observations to skip between each fold.}
}
\value{
A tibble with accuracy statistics.
}
\description{
Compute forecasts with a rolling origin and return accuracy statistics
}
